I/O thread Reference implementation.
This commit adds a reference implementation for the unix frontend to show how an I/O thread could be implemented.

Note that the global lock is still in place to allow for easily disabling the I/O thread. Performance may suffer, but the unix frontend is proof-of-concept only, anyway. The I/O thread can be disabled by commenting out the #define USE_IOTHREAD line in unix/main.cc

Access to guest RAM is bypassing the I/O thread because it is synchronized by the operating system.